requirements.txt
amqp==5.0.8
aniso8601==7.0.0
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
branca==0.4.2
CacheControl==0.12.10
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
decorator==5.1.0
Django==3.2.10
django-cors-headers==3.10.1
django-filter==21.1
easy-thumbnails==2.8
firebase-admin==5.2.0
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
msgpack==1.0.3
numpy==1.21.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfcm==1.5.4
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
ratelim==0.1.6
reportlab==3.6.3
requests==2.26.0
rsa==4.8
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
text-unidecode==1.3
tinycss2==1.1.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
